Moved HeadAction to play.core.actions
Fixes #2861, ensures that there's no play.api.controllers package to
interfere with controllers imports.
