+++ framework/src/play/src/main/scala/play/core/actions/HeadAction.scala
@@ -2,7 +2,7 @@
  * Copyright (C) 2009-2013 Typesafe Inc. <http://www.typesafe.com>
  */
 
-package play.api.controllers
+package play.core.actions
 
 import play.api.mvc._
 import play.api.libs.iteratee._
